# Conflicts:
#	t1/SEP3BlazorT1Client/SEP3BlazorT1Client/bin/Debug/net5.0/SEP3BlazorT1Client.Views.dll
#	t1/SEP3BlazorT1Client/SEP3BlazorT1Client/bin/Debug/net5.0/SEP3BlazorT1Client.Views.pdb
#	t1/SEP3BlazorT1Client/SEP3BlazorT1Client/bin/Debug/net5.0/SEP3BlazorT1Client.dll
#	t1/SEP3BlazorT1Client/SEP3BlazorT1Client/bin/Debug/net5.0/SEP3BlazorT1Client.pdb
#	t1/SEP3BlazorT1Client/SEP3BlazorT1Client/obj/Debug/net5.0/Razor/App.razor.g.cs
#	t1/SEP3BlazorT1Client/SEP3BlazorT1Client/obj/Debug/net5.0/Razor/Pages/RegisterResidence/RegisterResidence.razor.g.cs
#	t1/SEP3BlazorT1Client/SEP3BlazorT1Client/obj/Debug/net5.0/RazorDeclaration/SEP3BlazorT1Client.dll
#	t1/SEP3BlazorT1Client/SEP3BlazorT1Client/obj/Debug/net5.0/RazorDeclaration/SEP3BlazorT1Client.pdb
#	t1/SEP3BlazorT1Client/SEP3BlazorT1Client/obj/Debug/net5.0/SEP3BlazorT1Client.Views.dll
#	t1/SEP3BlazorT1Client/SEP3BlazorT1Client/obj/Debug/net5.0/SEP3BlazorT1Client.Views.pdb
#	t1/SEP3BlazorT1Client/SEP3BlazorT1Client/obj/Debug/net5.0/SEP3BlazorT1Client.dll
#	t1/SEP3BlazorT1Client/SEP3BlazorT1Client/obj/Debug/net5.0/SEP3BlazorT1Client.pdb

+++ t2/SEP3T2API/SEP3T2API/SEP3T2GraphQL/Models/Residence.cs
@@ -8,9 +8,9 @@ namespace SEP3T2GraphQL.Models
         public Address Address { get; set; }
         public string Description { get; set; }
         public string Type { get; set; }
-        public decimal AverageRating { get; set; }
+        public double AverageRating { get; set; }
         public bool IsAvailable { get; set; }
-        public decimal PricePerNight { get; set; }
+        public double PricePerNight { get; set; }
         public IList<Rule> Rules { get; set; } = new List<Rule>();
         public IList<Facility> Facilities { get; set; } = new List<Facility>(); 
     }
